cleanup (EventStatCard): move to `~/components/Base`
