+++ components/Base/EventStatCard/EventStatCardSkeleton.vue
